Squashed 'libraries/ZWidget/' changes from cecd34301d..924b8e0e78
924b8e0e78 Fix scroll size issue
f0344ed99b Add option to hide scrollbar when no scrolling is to be had
f7b01d6625 Remove stray sdl
214d098353 Theme role improvements
82d0431ca5 Textedit highlight now follows theme
373ab3f4ca Added hex Colorf alias
5ea6d5a62a Reduced duplicate code
a92e045b18 Simplified theme
d076c9991c Expanded example
e4556c97df Named colors
6d9300c742 Fix COM thread error
6499e5c523 Add null pointer checks
f9d5724210 Fix x11 backend locking up Implement raw mouse input support (XInput 2.0) Fix x11 DPI scaling
4a86fb5056 XInitThreads must be called when using vulkan
8095090774 Add DPI scale to X11 backend
02fb328172 Mute a bunch of warnings
f13c745a63 Fix typo
5689bf5edd Merge in changes from SurrealEngine Use C++20 Make SDL2 optional

git-subtree-dir: libraries/ZWidget
git-subtree-split: 924b8e0e786093708c0d73419a347f8e6ffe55d6

+++ src/core/utf8reader.cpp
@@ -71,7 +71,7 @@ unsigned int UTF8Reader::character()
 	else
 	{
 		unsigned int ucs4 = (data[current_position] & bitmask_leadbyte_for_utf8[trailing_bytes]);
-		for (std::string::size_type i = 0; i < trailing_bytes; i++)
+		for (std::string::size_type i = 0; i < (size_t)trailing_bytes; i++)
 		{
 			if ((data[current_position + 1 + i] & 0xC0) == 0x80)
 				ucs4 = (ucs4 << 6) + (data[current_position + 1 + i] & 0x3f);
@@ -94,7 +94,7 @@ std::string::size_type UTF8Reader::char_length()
 		if (current_position + 1 + trailing_bytes > length)
 			return 1;
 
-		for (std::string::size_type i = 0; i < trailing_bytes; i++)
+		for (std::string::size_type i = 0; i < (size_t)trailing_bytes; i++)
 		{
 			if ((data[current_position + 1 + i] & 0xC0) != 0x80)
 				return 1;
@@ -136,7 +136,7 @@ void UTF8Reader::move_to_leadbyte()
 			lead_position--;
 
 		int trailing_bytes = trailing_bytes_for_utf8[data[lead_position]];
-		if (lead_position + trailing_bytes >= current_position)
+		if (lead_position + trailing_bytes >= (int)current_position)
 			current_position = lead_position;
 	}
